+++ RELEASE_NOTES.md
@@ -43,6 +43,8 @@ This is now fixed, but might result in build previously working now fail due to 
 backports:
 - Support for implicit dependencies in target definitions
 - Add tycho-baseline:check-dependencies mojo
+- Add tycho-cleancode plugin
+
 
 ## 4.0.10